pyproject.toml
[project]
name = "shoppingai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "colorama>=0.4.6",
    "fastapi[standard]>=0.115.13",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "pydantic-settings>=2.9.1",
    "redis>=6.2.0",
    "streamlit>=1.46.0",
    "tortoise-orm[asyncpg]>=0.25.1",
]
